refactor: forbid admin to access cart

+++ app/Http/Controllers/CartController.php
@@ -9,14 +9,18 @@ use Illuminate\Support\Facades\Auth;
 
 class CartController extends Controller
 {
-    public function index()
+    public function index(Request $request)
     {
+        abort_if($request->user()->is_admin, 403);
+
         $cartItems = Auth::user()->carts()->with('product')->get();
         return view('user.cart', compact('cartItems'));
     }
 
-    public function store(Product $product)
+    public function store(Request $request, Product $product)
     {
+        abort_if($request->user()->is_admin, 403);
+
         $cart = Cart::firstOrNew([
             'user_id' => Auth::id(),
             'product_id' => $product->id,
@@ -28,8 +32,10 @@ class CartController extends Controller
         return back()->with('success', 'Produk ditambahkan ke keranjang');
     }
 
-    public function destroy(Cart $cart)
+    public function destroy(Request $request, Cart $cart)
     {
+        abort_if($request->user()->is_admin, 403);
+
         $cart->delete();
         return back()->with('success', 'Produk dihapus dari keranjang');
     }
